sysfs: Move new epoll code from fs into sysfs
I don't think it'll be used elsewhere after all. If it ever become the
case, we can revisit. But for now, it's better to not export
unnecessarily.

No functional change, just a move of unused code.
