+++ host/sysfs/fs_linux_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed under the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-package fs
+package sysfs
 
 import (
 	"io/ioutil"
